services/auth-service/requirements.txt
fastapi
uvicorn[standard]
redis
bcrypt
PyJWT==2.8.0
pydantic

